% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents}
\alias{accidents}
\title{Accidents Data}
\format{
A data frame with 192 monthly observations of 4 variables
\describe{
\item{children_accidents}{number of children aged 6 to 10 years seriously injured or
killed}
\item{children_exposure}{estimated number of children exposed}
\item{seniors_accidents}{senior persons above 65 seriously injured or killed}
\item{seniors_exposure}{estimated number of seniors exposed}
}
}
\source{
TBD
}
\usage{
accidents
}
\description{
Seriously injured or killed pedestrians per month in Linz, Austria, from
January 1987 to December 2002.
}
\keyword{datasets}
