% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{Dirichlet}
\alias{Dirichlet}
\alias{rdirichlet}
\title{The Dirichlet Distribution}
\usage{
rdirichlet(ndraws, alpha)
}
\arguments{
\item{ndraws}{number of observations.}

\item{alpha}{positive parameter vector.}
}
\value{
\code{rdirichlet} generates random deviates.
}
\description{
Random generation for the Dirichlet distribution with parameter
vector alpha.
}
