% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eyetracking}
\alias{eyetracking}
\title{Eye Tracking Data}
\format{
A data frame with 101 observations of 1 variable
\describe{
\item{anomalies}{number of eye tracking anomalies}
}
}
\source{
Escobar & West (1998)
}
\usage{
eyetracking
}
\description{
Count data on eye tracking anomalies in 101 schizophrenic patients,
sorted by size
}
\keyword{datasets}
