% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{GDP data}
\format{
A sequence of length 313
}
\source{
U.S. Bureau of Economic Analysis, Gross Domestic Product (GDP),
retrieved from FRED, Federal Reserve Bank of St. Louis;
\url{https://fred.stlouisfed.org/series/GDP}, July 24, 2025.
}
\usage{
gdp
}
\description{
Gross domestic product of the USA. Quarterly data from 1947Q1 to 2025Q1.
}
\details{
Gross domestic product (GDP), the featured measure of U.S. output,
is the market value of the goods and services produced by labor and property
located in the United States. The units are Billions of Dollars,
Seasonally Adjusted Annual Rate.
}
\keyword{datasets}
