% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflation}
\alias{inflation}
\title{Inflation Data}
\format{
A sequence of length 342.
}
\source{
European Central Bank Data Portal, retrieved from
\url{https://data.ecb.europa.eu/data/datasets/ICP/ICP.M.U2.N.000000.4.ANR},
July 25, 2025.
}
\usage{
inflation
}
\description{
Harmonized Index of Consumer Prices (HICP) in the Euro area (changing
composition). Monthly data from January 1997 to June 2025.
}
\keyword{datasets}
