% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R
\name{InvGammaDist}
\alias{InvGammaDist}
\alias{dinvgamma}
\alias{pinvgamma}
\alias{qinvgamma}
\alias{rinvgamma}
\title{Density, distribution function, quantile function and random generation for
the inverse gamma distribution with parameters a and b.}
\usage{
dinvgamma(x, a, b, log = FALSE)

pinvgamma(q, a, b)

qinvgamma(p, a, b)

rinvgamma(ndraws, a, b)
}
\arguments{
\item{x}{vector of quantiles.}

\item{a}{positive parameter.}

\item{b}{positive parameter.}

\item{log}{logical; if \code{TRUE}, densities are returned as logarithms.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{ndraws}{number of observations.}
}
\value{
\code{dinvgamma} gives the density, \code{pinvgamma} gives the
distribution function, \code{qinvgamma} gives the quantile function, and
\code{rinvgamma} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for
the inverse gamma distribution with parameters a and b.
}
