% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labor}
\alias{labor}
\title{Labor Market Data}
\format{
A data frame with 4376 observations (persons) of 6 variables
\describe{
\item{age}{age of the person in the year 1986}
\item{unemp97}{logical, TRUE if the person was unemployed in 1997}
\item{unemp98}{logical, TRUE if the person was unemployed in 1998}
\item{changeempl}{number of times the person changed employer from 1986 until 1998}
\item{female}{logical, TRUE if the person is female}
\item{wcollar}{logical, TRUE if the person was classified as white collar employee}
}
}
\source{
Weber (2001)
}
\usage{
labor
}
\description{
Labor market data of Austrian employees. The data stem from 1986--1998.
}
\keyword{datasets}
