% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{Logsumexp}
\alias{Logsumexp}
\alias{logsumexp}
\alias{logmeanexp}
\title{The Log-sum-exp function and the Log-mean-exp function}
\usage{
logsumexp(x)

logmeanexp(x)
}
\arguments{
\item{x}{sequence of real numbers}
}
\value{
a single real number
}
\description{
Numerically stable variant to compute the log of the sum (mean) of exponentials
}
