% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movies}
\alias{movies}
\title{Movie Open Box Office Data}
\format{
A data frame with 94 observations of 29 variables
\describe{
\item{title}{The title of the movie (anonymized)}
\item{...}{etc. etc.}
}
}
\source{
TBD: \url{https://doi.org/10.7910/DVN/JS3AVM}
}
\usage{
movies
}
\description{
TBD
}
\keyword{datasets}
