% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newcars}
\alias{newcars}
\title{New Cars Registration Data}
\format{
A monthly time series object with 313 observations.
}
\source{
Statistik Austria open.data, retrieved from
\url{https://data.statistik.gv.at/web/meta.jsp?dataset=OGD_fkfzul0759_OD_PkwNZL_1},
March 5, 2026.
}
\usage{
newcars
}
\description{
Registrations counts of new passenger cars in Austria.
Monthly time series data from January 2000 to January 2026.
}
\keyword{datasets}
