% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentt.R
\name{StudTDist}
\alias{StudTDist}
\alias{dstudt}
\alias{pstudt}
\alias{qstudt}
\alias{rstudt}
\title{The Non-standardized Student t Distribution}
\usage{
dstudt(x, location = 0, scale = 1, df, log = FALSE)

pstudt(x, location = 0, scale = 1, df)

qstudt(p, location = 0, scale = 1, df)

rstudt(n, location = 0, scale = 1, df)
}
\arguments{
\item{x}{vector of quantiles.}

\item{location}{parameter.}

\item{scale}{positive parameter.}

\item{df}{positive parameter.}

\item{log}{logical; if \code{TRUE}, densities are returned as logarithms.}

\item{p}{vector of probabilities.}

\item{q}{vector of quantiles.}

\item{ndraws}{number of observations.}
}
\value{
\code{dstudt} gives the density, \code{pstudt} gives the
distribution function, \code{qstudt} gives the quantile function, and
\code{rstudt} generates random deviates.
}
\description{
Density, distribution function, quantile function and random generation for
the location-scale-transformed Student t distribution.
}
\details{
The non-standardized Student t distribution with location \eqn{\mu}, scale
\eqn{\tau}, and df \eqn{\nu > 2} has mean \eqn{\mu} and variance
\eqn{\tau^2 \frac{\nu}{\nu-2}}.
}
