% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{words}
\alias{words}
\title{Words Data}
\format{
A data frame with 100 observations of 2 variables
\describe{
\item{word}{an English word (lemma)}
\item{frequency}{number of occurrences}
}
}
\source{
Sourced from \url{https://www.wordfrequency.info} on March 20, 2025
}
\usage{
words
}
\description{
List of the 100 most common English lemmas (i.e., words whose different
forms have been grouped together), according to the Corpus of Contemporary
American English (COCA).
}
\keyword{datasets}
