% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eyetracking}
\alias{eyetracking}
\title{Eye Tracking Data}
\format{
An object of class \code{integer} of length 192.
}
\source{
Escobar & West (1998)
}
\usage{
eyetracking
}
\description{
Count data on eye tracking anomalies in 101 schizophrenic patients
}
\keyword{datasets}
